requirements.txt
pdfplumber
openai
faiss-cpu
numpy
python-dotenv
